Ready

+++ src/main/java/com/example/spring_rest/dao/UserDAOImpl.java
@@ -1,23 +1,23 @@
 package com.example.spring_rest.dao;
 
 import com.example.spring_rest.entity.User;
-import java.util.List;
-import javax.persistence.EntityManager;
+import lombok.RequiredArgsConstructor;
 import org.hibernate.Session;
 import org.hibernate.query.Query;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
+import javax.persistence.EntityManager;
+import java.util.List;
+
+/**
+ * Класс, который ответственен за работу с Базой данных
+ */
 @Repository
-// Класс, который ответственен за работу с Базой данных
+@RequiredArgsConstructor(onConstructor = @__(@Autowired))
 public class UserDAOImpl implements UserDAO {
 
     private final EntityManager entityManager;
-
-    @Autowired
-    public UserDAOImpl(EntityManager entityManager) {
-        this.entityManager = entityManager;
-    }
 
     @Override
     public List<User> getAllUsers() {
@@ -41,8 +41,7 @@ public class UserDAOImpl implements UserDAO {
     @Override
     public void deleteUser(int id) {
         Session session = entityManager.unwrap(Session.class);
-        Query<?> query = session.createQuery("delete from User " +
-                "where id =:userId");
+        Query<?> query = session.createQuery("delete from User where id =:userId");
         query.setParameter("userId", id);
         query.executeUpdate();
     }
